Effect handling methods have been moved from the Entity class to a seperate EffectHandler class for encapsulations sake. EffectHandler assignment is prone to breaking, so should monitor.
